Added intellij, datagrip, easyssh, google chrome browser, cromedriver

+++ manjaro19/basic/remove_canonical_stuff.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -e
-# We do not want Canonical stuff)
-sudo systemctl stop snapd.socket
-sudo systemctl disable snapd.socket
-sudo pacman -Rns --noconfirm snapd
+if command -v snap;
+then
+	# We do not want Canonical stuff)
+	sudo systemctl stop snapd.socket
+	sudo systemctl disable snapd.socket
+	sudo pacman -Rns --noconfirm snapd
+else
+	printf "${MSGCOLOR}Canonical stuff allready removed${NOCOLOR}\n"
+fi
